requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
asyncpg
pydantic